Выводит шапку
