+++ src/components/form.js
@@ -1,6 +1,6 @@
 import {createElement} from '../utils.js';
 
-export const createFormTemplate = () => {
+const createFormTemplate = () => {
   return (
     `<form class="trip-events__item  event  event--edit" action="#" method="post">
       <header class="event__header">
