+++ src/components/sorting.js
@@ -1,6 +1,6 @@
 import {createElement} from '../utils.js';
 
-export const createSortingTemplate = () => {
+const createSortingTemplate = () => {
   return (
     `<form class="trip-events__trip-sort  trip-sort" action="#" method="get">
       <span class="trip-sort__item  trip-sort__item--day">Day</span>
